+++ NEWS.md
@@ -1,5 +1,6 @@
-
 # gtsummary (development version)
+
+* Option to exclude some variables from testing when using `add_p()` (#164)
 
 * Updates after {gt} package renamed `cells_style()` to `cell_text()` (#78)